pyproject.toml
[project]
name = "gundi-dlq"
version = "0.2.1"
description = "Utility to re-process messages from Gundi dead-letter queues"
authors = [
    {name = "Mariano Martinez Grasso",email = "marianomartinez84@gmail.com"}
]
license = "Apache-2.0"
readme = "README.md"
packages = [
    { include = "gundi_dlq_processor" }
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "gcloud-aio-pubsub (>=6.0.1,<7.0.0)",
    "click (>=8.1.8,<9.0.0)"
]
[tool.poetry.scripts]
gundi-dlq = "gundi_dlq:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
